fixed bug

+++ offchain_actors/src/operator.rs
@@ -451,7 +451,7 @@ async fn main_loop(
 
                 let unwrapped_just = justification.unwrap().unwrap();
 
-                if justification_to_process.is_none() && unwrapped_just.commit.target_number >= last_processed_block_num.unwrap() + MAX_NUM_HEADERS_PER_STEP {
+                if justification_to_process.is_none() && unwrapped_just.commit.target_number >= last_processed_block_num.unwrap() + (MAX_NUM_HEADERS_PER_STEP as u32) {
                     println!("Saving justification for block number: {:?}", unwrapped_just.commit.target_number);
                     justification_to_process = Some(unwrapped_just);
                 }
